CHANGED:    django app for shopping cart
