+++ backend/urls.py
@@ -24,7 +24,8 @@ urlpatterns = [
     path('admin/', admin.site.urls),
     path('api/store/', include('store.urls')),
     path('api/blog/', include('blog.urls')),
-    path('api/account/', include('tep_user.urls'))
+    path('api/account/', include('tep_user.urls')),
+    path('api/cart/', include('cart.urls'))
 ]
