+++ cart/urls.py
@@ -1,0 +1,10 @@
+from django.urls import path, include
+from rest_framework.routers import DefaultRouter
+from .views import CartViewSet
+
+router = DefaultRouter()
+router.register('', CartViewSet, basename='cart')
+
+urlpatterns = [
+    path('', include(router.urls)),
+]
