+++ tep_user/admin.py
@@ -1,7 +1,6 @@
 from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin
-
-from tep_user.models import TEPUser, Cart
+from tep_user.models import TEPUser
 
 
 @admin.register(TEPUser)
@@ -11,14 +10,3 @@ class TEPUserAdmin(UserAdmin):
     search_fields = ('email', 'first_name', 'last_name')
     ordering = ('email',)
 
-
-@admin.register(Cart)
-class CartAdmin(admin.ModelAdmin):
-    list_display = ('id', 'get_user_email')
-    search_fields = ('id', 'tep_user__email',)
-    filter_horizontal = ('product_variants',)
-
-    def get_user_email(self, obj):
-        return obj.tep_user.email
-
-    get_user_email.short_description = 'User Email'
